Sono state apportate le seguenti modifiche: - Aumentate le prestazioni di precaricamento - Eseguito refactoring del componente TopNav - Eseguito refactoring del componente RichiestaPreventivo
