+++ index.html
@@ -23,7 +23,7 @@
 
 <body>
   <div id="app"></div>
-  <script type="module" src="/src/main.js"></script>
+  <script defer="defer" type="module" src="/src/main.js"></script>
 </body>
 
 </html>
