+++ src/plugins/index.js
@@ -10,7 +10,7 @@ import vuetify from './vuetify'
 import router from '../router'
 
 export function registerPlugins (app) {
-  loadFonts()
+  loadFonts();
   app
     .use(vuetify)
     .use(router)
